delivery_dashboard/requirements.txt
flask
flask-socketio
kafka-python